# 8331114: Further improve performance of MethodTypeDesc::descriptorString