--- a/src/java.base/share/classes/java/lang/constant/ClassDesc.java
+++ b/src/java.base/share/classes/java/lang/constant/ClassDesc.java
@@ -160,7 +160,7 @@ public sealed interface ClassDesc
     static ClassDesc ofDescriptor(String descriptor) {
         // implicit null-check
         return (descriptor.length() == 1)
-               ? Wrapper.forPrimitiveType(descriptor.charAt(0)).primitiveClassDescriptor()
+               ? Wrapper.forPrimitiveType(descriptor.charAt(0)).classDescriptor()
                // will throw IAE on descriptor.length == 0 or if array dimensions too long
                : new ReferenceClassDescImpl(descriptor);
     }