--- a/src/java.base/share/classes/java/lang/constant/MethodTypeDesc.java
+++ b/src/java.base/share/classes/java/lang/constant/MethodTypeDesc.java
@@ -204,15 +204,11 @@ public sealed interface MethodTypeDesc
      * @return the human-readable descriptor for this method type
      */
     default String displayDescriptor() {
-        int count = parameterCount();
-        StringBuilder sb = new StringBuilder(24).append('(');
-        for (int i = 0; i < count; i++) {
-            if (i > 0) {
-                sb.append(',');
-            }
-            sb.append(parameterType(i).displayName());
-        }
-        return sb.append(')').append(returnType().displayName()).toString();
+        return String.format("(%s)%s",
+                             Stream.of(parameterArray())
+                                   .map(ClassDesc::displayName)
+                                   .collect(Collectors.joining(",")),
+                             returnType().displayName());
     }
 
     /**