--- a/src/java.base/share/classes/java/lang/constant/MethodTypeDescImpl.java
+++ b/src/java.base/share/classes/java/lang/constant/MethodTypeDescImpl.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, 2023, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2018, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -178,7 +178,7 @@ final class MethodTypeDescImpl implements MethodTypeDesc {
         for (ClassDesc argType : argTypes) {
             sb.append(argType.descriptorString());
         }
-        desc = sb.append(')').append(returnType().descriptorString()).toString();
+        desc = sb.append(')').append(returnType.descriptorString()).toString();
         cachedDescriptorString = desc;
         return desc;
     }