--- a/test/micro/org/openjdk/bench/java/lang/constant/MethodTypeDescFactories.java
+++ b/test/micro/org/openjdk/bench/java/lang/constant/MethodTypeDescFactories.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2023, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it