When analyzing (startup) performance of the Classfile API I found this opportunity to further improve `MethodTypeDescImpl::descriptorString`.

Performance improves across the board in existing microbenchmarks:
```
Name                                                                 (descString) Cnt   Base   Error    Test   Error  Unit  Change
MethodTypeDescFactories.descriptorString  (Ljava/lang/Object;Ljava/lang/String;)I   6 55,179 ± 2,027  32,920 ± 1,189 ns/op   1,68x (p = 0,000*)
MethodTypeDescFactories.descriptorString                                      ()V   6 17,689 ± 1,871  11,060 ± 0,331 ns/op   1,60x (p = 0,000*)
MethodTypeDescFactories.descriptorString ([IJLjava/lang/String;Z)Ljava/util/List;   6 86,627 ± 1,646  41,035 ± 0,636 ns/op   2,11x (p = 0,000*)
MethodTypeDescFactories.descriptorString                    ()[Ljava/lang/String;   6 18,305 ± 1,974  13,110 ± 0,089 ns/op   1,40x (p = 0,000*)
  * = significant
```

The improvement is even more pronounced when running with `-Xint`, which is relevant for reducing startup overheads of early ClassFile API use:
```
Name                                                                 (descString) Cnt     Base      Error      Test     Error  Unit  Change
MethodTypeDescFactories.descriptorString  (Ljava/lang/Object;Ljava/lang/String;)I   6 5122,061 ±   81,335  2626,481 ± 101,466 ns/op   1,95x (p = 0,000*)
MethodTypeDescFactories.descriptorString                                      ()V   6 3481,316 ±  258,904  1489,267 ±  15,506 ns/op   2,34x (p = 0,000*)
MethodTypeDescFactories.descriptorString ([IJLjava/lang/String;Z)Ljava/util/List;   6 7741,081 ± 1628,244  3281,778 ±  41,892 ns/op   2,36x (p = 0,000*)
MethodTypeDescFactories.descriptorString                    ()[Ljava/lang/String;   6 3677,803 ±   63,432  1495,291 ±   8,995 ns/op   2,46x (p = 0,000*)
  * = significant
  ```
  
 I also applied similar approach to `MethodTypeDesc::displayDescriptor`: while not performance sensitive I think these are so inter-related that it makes sense to implement them in a similar fashion.

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8331114](https://bugs.openjdk.org/browse/JDK-8331114): Further improve performance of MethodTypeDesc::descriptorString (**Enhancement** - P4)


### Reviewers
 * [Mandy Chung](https://openjdk.org/census#mchung) (@mlchung - **Reviewer**)
 * [Chen Liang](https://openjdk.org/census#liach) (@liach - Author) ⚠️ Review applies to [7622f8f4](https://git.openjdk.org/jdk/pull/18945/files/7622f8f4dd8b4304d6334655ea56c881ee9fd249)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/18945/head:pull/18945` \
`$ git checkout pull/18945`

Update a local copy of the PR: \
`$ git checkout pull/18945` \
`$ git pull https://git.openjdk.org/jdk.git pull/18945/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 18945`

View PR using the GUI difftool: \
`$ git pr show -t 18945`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/18945.diff">https://git.openjdk.org/jdk/pull/18945.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/18945#issuecomment-2076788647)