@liach 2024-04-25
Do we have any research on the average length of method descriptor strings? I wonder if manual pre-allocation works better (iterating all descriptor strings, allocate the sum of their sizes plus 2 (for parentheses), as descriptor strings won't be re-calculated after initial allocation.)
Especially in case of user code, as many user classes have very long package names that can easily make the string much longer than the 24-char default.