@liach 2024-04-25
What if we replace `24` with a precalculated value:
```java
int size = 2 + returnType().descriptorString().length();
for (var param : argTypes)
    size += param.descriptorString().length();
```
(Would be even better if we can just trust the internal array to avoid copy allocation)