@cl4es 2024-04-25
Correction: I ran the wrong benchmark for that long descriptor test! 🤦🏽 On the `descriptorString` benchmark there _is_ a regression with my patch on longer descriptors, -30% on that example. Since the strings are all readily available pre-calculating the size is cheap enough to make sense. This makes the patch more or less neutral on large descriptors, extends the gain on few-arg methods, but reduces the win on cases where there are no args (due an extra branch et.c.). On balance this seems like an improvement:

```
Name                                                                 (descString) Cnt    Base   Error    Test    Error  Unit  Change
MethodTypeDescFactories.descriptorString  (Ljava/lang/Object;Ljava/lang/String;)I   6  43,180 ± 0,575   30,547 ± 0,387 ns/op   1,41x (p = 0,000*)
MethodTypeDescFactories.descriptorString                                      ()V   6  20,717 ± 0,128   17,233 ± 0,242 ns/op   1,20x (p = 0,000*)
MethodTypeDescFactories.descriptorString ([IJLjava/lang/String;Z)Ljava/util/List;   6  89,834 ± 1,951   40,149 ± 0,326 ns/op   2,24x (p = 0,000*)
MethodTypeDescFactories.descriptorString                    ()[Ljava/lang/String;   6  21,438 ± 0,490   15,937 ± 0,263 ns/op   1,35x (p = 0,000*)
MethodTypeDescFactories.descriptorString                                 (..IIJ)V   6  52,429 ± 0,117   49,396 ± 0,153 ns/op   1,06x (p = 0,000*)
MethodTypeDescFactories.descriptorString                 (.....................).   6 178,205 ± 0,518  176,856 ± 5,577 ns/op   1,01x (p = 0,158 )
  * = significant
```

`.` is illegal in descriptor strings I used this as wildcard to be replaced with the descriptor string for the benchmark class as a ways to produce very large descriptors.

Since this strays a bit from what's doable in `MethodTypeDesc::displayDescriptor` I'm reverting those changes and focus this PR on the relevant piece of machinery.