@liach 2024-04-25
For the precise-length approach, do you have the results? I guess if we can avoid copying the stringbuilder array, we can make this even faster.