@cl4es 2024-04-26
> For the precise-length approach, do you have the results? I guess if we can avoid copying the stringbuilder array, we can make this even faster.

Not sure what you mean - the latest performance numbers I posted are for the precise length version (15c8d39) if that was unclear.